requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.62.0
jpype1>=1.4.0